backend/pyproject.toml
[project]
name = "cpt208"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "ragflow-sdk>=0.24.0",
    "uvicorn>=0.44.0",
]
